Update neo 3.6 (#758) * Fix safe assert * ASSERTMSG AND ABORTMSG * Fix one UT * Fix more uts * Fix event deploy * Manifest fix --------- Co-authored-by: Jimmy <jinghui@wayne.edu>

+++ tests/Neo.SmartContract.Framework.UnitTests/TestClasses/Contract_Contract.cs
@@ -27,7 +27,7 @@ namespace Neo.SmartContract.Framework.UnitTests.TestClasses
 
         public static int GetCallFlags()
         {
-            return Contract.GetCallFlags();
+            return (int)Contract.GetCallFlags();
         }
 
         public static UInt160 CreateStandardAccount(Cryptography.ECC.ECPoint pubKey)
